将哨兵工程从AC5编译器移植到AC6编译器

+++ lowerComputer/application/INS_task.c
@@ -236,7 +236,7 @@ void IMU_Temperature_Ctrl(void)
 {
     pid_calculate(&TempCtrl, BMI088.Temperature, RefTemp);
 
-    TIM_Set_PWM(&htim10, TIM_CHANNEL_1, fp32_constrain(fp32_rounding(TempCtrl.Output), 0, UINT32_MAX));
+    TIM_Set_PWM(&htim10, TIM_CHANNEL_1, fp32_constrain(fp32_rounding(TempCtrl.Output), 0, (fp32)UINT32_MAX));
 }
 
 //------------------------------------functions below are not used in this demo-------------------------------------------------
